% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportData.R
\name{ClipQuakes}
\alias{ClipQuakes}
\title{Select earthquake sover area}
\usage{
ClipQuakes(quakes, area)
}
\arguments{
\item{quakes}{data.table with earthquakes}

\item{area}{SpatialPolygons}
}
\description{
Select earthquake sover area
}

