% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportData.R
\name{Earthquakes}
\alias{Earthquakes}
\title{Loads the KNMI earthquake catalogue}
\usage{
Earthquakes(type = "induced", area = NULL, period = NULL, path = "")
}
\arguments{
\item{type}{Type of catalogue c('induced', 'tectonic')}

\item{area}{Inheriting from spatial polygon}

\item{period}{Either numeric, timeBased or ISO-8601 style (see \code{\link[xts]{.subset.xts}})}

\item{path}{for saving data}
}
\value{
data.table with rows being the single events
}
\description{
Loads the KNMI earthquake catalogue
}
\examples{
Earthquakes("induced", Groningen, "1990/2016")

}

