% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaData.R
\docType{data}
\name{EarthquakesBoundaryBox}
\alias{EarthquakesBoundaryBox}
\title{Earthquakes maximal domain}
\format{bbox i.e. 2*2 matrix}
\usage{
EarthquakesBoundaryBox
}
\description{
Earthquakes maximal domain
}
\keyword{datasets}

