% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOBS.R
\name{EOBS}
\alias{EOBS}
\title{Imports EOBS data (template for transition to WCS)}
\usage{
EOBS(variable, geoIdentifier, period, path, basegrid, regrid = FALSE,
  resx = 0.25, resy = 0.25)
}
\arguments{
\item{variable}{String(s) identifying variable(s)}

\item{geoIdentifier}{Either string(s) to identify stations,
SpatialPoint(s) for closest station/grid box,
bbox or extending SpatialPolygon for everything inside area}

\item{period}{Either numeric, timeBased or ISO-8601 style (see \code{\link[xts]{.subset.xts}})}

\item{path}{for file download (if set to NULL data are always downloaded but not saved)}

\item{basegrid}{String identifying the basegrid}

\item{regrid}{Boolean should it be regridded through WCS}

\item{resx}{x-Resolution for regridding (default 0.25)}

\item{resy}{y-Resolution for regridding (default 0.25)}
}
\description{
Imports EOBS data (template for transition to WCS)
}

