% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnmiDataClass.R
\name{getInformation}
\alias{Call}
\alias{Call.default}
\alias{Citation}
\alias{Description}
\alias{License}
\alias{MetaData}
\alias{getInformation}
\title{Information about KnmiData object}
\usage{
Call(x, ...)

\method{Call}{default}(x, ...)

Description(x, ...)

License(x, ...)

Citation(x, ...)

MetaData(x, ...)
}
\arguments{
\item{x}{KnmiData object}

\item{...}{Additional parameters (deprecated)}
}
\description{
Information about KnmiData object
}
\examples{
data <- Earthquakes("induced", NULL, "2016-04-01/2016-05-18", path = NULL)
Call(data)
Description(data)
License(data)
Citation(data)
MetaData(data)
}

