% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaData.R
\docType{data}
\name{Groningen}
\alias{Groningen}
\title{Groningen reservoir boundaries}
\format{SpatialPolygons}
\source{
KNMI
}
\usage{
Groningen
}
\description{
Groningen reservoir boundaries
}
\keyword{datasets}
