% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportData.R
\name{HomogenizedPrecipitation}
\alias{HomogenizedPrecipitation}
\title{Obtains Brandsma series for specified station and period}
\usage{
HomogenizedPrecipitation(stationId, periodStart = 1910)
}
\arguments{
\item{stationId}{Integer specifying the KNMI station number}

\item{periodStart}{Either 1910 or 1951}
}
\value{
data.table with date and precipitation values
}
\description{
The data are downloaded from the Climate Explorer.
}

