% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportData.R
\name{HomogenPrecip}
\alias{HomogenPrecip}
\title{Obtains homogenized precipitation data}
\usage{
HomogenPrecip(location, period, whichSet = "automatic")
}
\arguments{
\item{location}{Either inheriting from spatial polygon or station number}

\item{period}{Either numeric, timeBased or ISO-8601 style (see \code{\link[xts]{.subset.xts}})}

\item{whichSet}{Which set should be used c(1910, 1951, "automatic")?}
}
\value{
data.table with date, stationId and precipitation values
}
\description{
8-8 daily precipitation measurements
}

