% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportData.R
\name{IsInArea}
\alias{IsInArea}
\title{Is point in area?}
\usage{
IsInArea(points, area)
}
\arguments{
\item{points}{SpatialPoints}

\item{area}{SpatialPolygons of SpatialPolygonsDataFrame}
}
\value{
logical indicating whether point is in area
}
\description{
Is point in area?
}
