% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knmi.R
\name{knmi_get_data}
\alias{knmi_get_data}
\title{R wrapper for knmi daily data api}
\usage{
knmi_get_data(start, end, vars, stns)
}
\arguments{
\item{start}{The start time}

\item{end}{The end time}

\item{vars}{The variables to query}

\item{stns}{The stations to query}
}
\value{
A data.frame with reponse data
}
\description{
R wrapper for knmi daily data api
}
