% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnmiDataClass.R
\name{KnmiData}
\alias{KnmiData}
\title{Generate a KNMI data object}
\usage{
KnmiData(data, call, type)
}
\arguments{
\item{data}{The raw data (can be different objects)}

\item{call}{Call object describing the creation of the data}

\item{type}{String describing what kind of data set}
}
\value{
An augmentated data object
}
\description{
The creation of the same data should be facilitated by the
 storage of call, knmiR package version and timeStamp
}
