% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knmiR.R
\docType{package}
\name{knmiR}
\alias{knmiR}
\alias{knmiR-package}
\title{knmiR: A package to import and KNMI data}
\description{
The package provides an R interface to some data sources of KNMI.
}
\details{
Available data sources:

Earthquake catalog via the function Earthquakes()

KIS data base via the function KIS() (internally only)
}
