% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOBS.R
\name{LocalImportEOBS}
\alias{LocalImportEOBS}
\title{Import EOBS data from local file}
\usage{
LocalImportEOBS(variable, filename, period = NULL, area = NULL,
  na.rm = TRUE)
}
\arguments{
\item{filename}{String containing the path to the ncdf file}
}
\description{
Import EOBS data from local file
}
\note{
Should be merged with \code{\link{importEOBS}}
}
