% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaData.R
\docType{data}
\name{stationMetaData}
\alias{stationMetaData}
\title{Homogenized precipitation station meta data}
\format{A data.table with 240 rows and 4 variables:
\describe{
  \item{stationId}{KNMI station id},
  \item{lat}{Latitude},
  \item{lon}{Longitude},
  \item{elev}{Elevation in meters}
  \item{stationName}{Name of the Station}
  \item{longRecord}{TRUE if record starts in 1910}}}
\source{
T. Brandsma \url{brandsma@knmi.nl}
}
\usage{
stationMetaData
}
\description{
A dataset containing the station meta data for the stations
used in Buishand et al. (2013)
}
\keyword{datasets}
