% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyArea.R
\name{StudyArea}
\alias{StudyArea}
\title{Create study area sp from coordinates}
\usage{
StudyArea(coords, bbox = FALSE)
}
\arguments{
\item{coords}{coordinate matrix row point, first col lon, second col lat}

\item{bbox}{boolean specifying wether a SpatialPolygon representing the bbox
of the coordinates should be returned}
}
\value{
SpatialPolygons object
}
\description{
Create study area sp from coordinates
}
